+++ src/RcppExports.cpp
@@ -6,58 +6,6 @@
 
 using namespace Rcpp;
 
-// cpp_pairwise_L2
-Rcpp::List cpp_pairwise_L2(arma::mat muA, arma::mat muB, arma::cube covA, arma::cube covB);
-RcppExport SEXP _maotai_cpp_pairwise_L2(SEXP muASEXP, SEXP muBSEXP, SEXP covASEXP, SEXP covBSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type muA(muASEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type muB(muBSEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type covA(covASEXP);
-    Rcpp::traits::input_parameter< arma::cube >::type covB(covBSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_pairwise_L2(muA, muB, covA, covB));
-    return rcpp_result_gen;
-END_RCPP
-}
-// dat2centers
-arma::vec dat2centers(arma::rowvec data, arma::mat& centers);
-RcppExport SEXP _maotai_dat2centers(SEXP dataSEXP, SEXP centersSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::rowvec >::type data(dataSEXP);
-    Rcpp::traits::input_parameter< arma::mat& >::type centers(centersSEXP);
-    rcpp_result_gen = Rcpp::wrap(dat2centers(data, centers));
-    return rcpp_result_gen;
-END_RCPP
-}
-// gradF
-arma::mat gradF(Function func, arma::mat xnow, double h);
-RcppExport SEXP _maotai_gradF(SEXP funcSEXP, SEXP xnowSEXP, SEXP hSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Function >::type func(funcSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type xnow(xnowSEXP);
-    Rcpp::traits::input_parameter< double >::type h(hSEXP);
-    rcpp_result_gen = Rcpp::wrap(gradF(func, xnow, h));
-    return rcpp_result_gen;
-END_RCPP
-}
-// solve_lyapunov
-arma::mat solve_lyapunov(arma::mat A, arma::mat B, arma::mat C);
-RcppExport SEXP _maotai_solve_lyapunov(SEXP ASEXP, SEXP BSEXP, SEXP CSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type A(ASEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type B(BSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type C(CSEXP);
-    rcpp_result_gen = Rcpp::wrap(solve_lyapunov(A, B, C));
-    return rcpp_result_gen;
-END_RCPP
-}
 // aux_shortestpath
 Rcpp::NumericMatrix aux_shortestpath(NumericMatrix& wmat);
 RcppExport SEXP _maotai_aux_shortestpath(SEXP wmatSEXP) {
@@ -66,19 +14,6 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix& >::type wmat(wmatSEXP);
     rcpp_result_gen = Rcpp::wrap(aux_shortestpath(wmat));
-    return rcpp_result_gen;
-END_RCPP
-}
-// cpp_sylvester
-arma::mat cpp_sylvester(arma::mat A, arma::mat B, arma::mat C);
-RcppExport SEXP _maotai_cpp_sylvester(SEXP ASEXP, SEXP BSEXP, SEXP CSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type A(ASEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type B(BSEXP);
-    Rcpp::traits::input_parameter< arma::mat >::type C(CSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_sylvester(A, B, C));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -96,6 +31,83 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< int >::type maxiter(maxiterSEXP);
     Rcpp::traits::input_parameter< double >::type eps(epsSEXP);
     rcpp_result_gen = Rcpp::wrap(cppsub_2007Wang(V0, mm, d, Spu, Stu, maxiter, eps));
+    return rcpp_result_gen;
+END_RCPP
+}
+// gradF
+arma::mat gradF(Function func, arma::mat xnow, double h);
+RcppExport SEXP _maotai_gradF(SEXP funcSEXP, SEXP xnowSEXP, SEXP hSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Function >::type func(funcSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type xnow(xnowSEXP);
+    Rcpp::traits::input_parameter< double >::type h(hSEXP);
+    rcpp_result_gen = Rcpp::wrap(gradF(func, xnow, h));
+    return rcpp_result_gen;
+END_RCPP
+}
+// dat2centers
+arma::vec dat2centers(arma::rowvec data, arma::mat& centers);
+RcppExport SEXP _maotai_dat2centers(SEXP dataSEXP, SEXP centersSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::rowvec >::type data(dataSEXP);
+    Rcpp::traits::input_parameter< arma::mat& >::type centers(centersSEXP);
+    rcpp_result_gen = Rcpp::wrap(dat2centers(data, centers));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_sylvester
+arma::mat cpp_sylvester(arma::mat A, arma::mat B, arma::mat C);
+RcppExport SEXP _maotai_cpp_sylvester(SEXP ASEXP, SEXP BSEXP, SEXP CSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type A(ASEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type B(BSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type C(CSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_sylvester(A, B, C));
+    return rcpp_result_gen;
+END_RCPP
+}
+// solve_lyapunov
+arma::mat solve_lyapunov(arma::mat A, arma::mat B, arma::mat C);
+RcppExport SEXP _maotai_solve_lyapunov(SEXP ASEXP, SEXP BSEXP, SEXP CSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type A(ASEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type B(BSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type C(CSEXP);
+    rcpp_result_gen = Rcpp::wrap(solve_lyapunov(A, B, C));
+    return rcpp_result_gen;
+END_RCPP
+}
+// cpp_pairwise_L2
+Rcpp::List cpp_pairwise_L2(arma::mat muA, arma::mat muB, arma::cube covA, arma::cube covB);
+RcppExport SEXP _maotai_cpp_pairwise_L2(SEXP muASEXP, SEXP muBSEXP, SEXP covASEXP, SEXP covBSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::mat >::type muA(muASEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type muB(muBSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type covA(covASEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type covB(covBSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_pairwise_L2(muA, muB, covA, covB));
+    return rcpp_result_gen;
+END_RCPP
+}
+// integrate_1d
+double integrate_1d(arma::vec& tseq, arma::vec& fval);
+RcppExport SEXP _maotai_integrate_1d(SEXP tseqSEXP, SEXP fvalSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec& >::type tseq(tseqSEXP);
+    Rcpp::traits::input_parameter< arma::vec& >::type fval(fvalSEXP);
+    rcpp_result_gen = Rcpp::wrap(integrate_1d(tseq, fval));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -155,13 +167,14 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_maotai_cpp_pairwise_L2", (DL_FUNC) &_maotai_cpp_pairwise_L2, 4},
-    {"_maotai_dat2centers", (DL_FUNC) &_maotai_dat2centers, 2},
-    {"_maotai_gradF", (DL_FUNC) &_maotai_gradF, 3},
-    {"_maotai_solve_lyapunov", (DL_FUNC) &_maotai_solve_lyapunov, 3},
     {"_maotai_aux_shortestpath", (DL_FUNC) &_maotai_aux_shortestpath, 1},
-    {"_maotai_cpp_sylvester", (DL_FUNC) &_maotai_cpp_sylvester, 3},
     {"_maotai_cppsub_2007Wang", (DL_FUNC) &_maotai_cppsub_2007Wang, 7},
+    {"_maotai_gradF", (DL_FUNC) &_maotai_gradF, 3},
+    {"_maotai_dat2centers", (DL_FUNC) &_maotai_dat2centers, 2},
+    {"_maotai_cpp_sylvester", (DL_FUNC) &_maotai_cpp_sylvester, 3},
+    {"_maotai_solve_lyapunov", (DL_FUNC) &_maotai_solve_lyapunov, 3},
+    {"_maotai_cpp_pairwise_L2", (DL_FUNC) &_maotai_cpp_pairwise_L2, 4},
+    {"_maotai_integrate_1d", (DL_FUNC) &_maotai_integrate_1d, 2},
     {"_maotai_eval_gaussian", (DL_FUNC) &_maotai_eval_gaussian, 3},
     {"_maotai_eval_gaussian_data", (DL_FUNC) &_maotai_eval_gaussian_data, 3},
     {"_maotai_eval_gmm_data", (DL_FUNC) &_maotai_eval_gmm_data, 4},